Essentials

Global State

Metal.device! — Function

device!(dev::MTLDevice)

Sets the Metal GPU device associated with the current Julia task.

Metal.MTL.devices — Function

devices()

Get an iterator for the compute devices.

Metal.current_device — Function

current_device()::MTLDevice

Return the Metal GPU device associated with the current Julia task.

Since all M-series systems currently only externally show a single GPU, this function effectively returns the only system GPU.

Metal.global_queue — Function

global_queue(dev::MTLDevice)::MTLCommandQueue

Return the Metal command queue associated with the current Julia thread.

Metal.synchronize — Function

synchronize(queue)

Wait for currently committed GPU work on this queue to finish.

Create a new MTLCommandBuffer from the global command queue, commit it to the queue, and simply wait for it to be completed. Since command buffers should execute in a First-In-First-Out manner, this synchronizes the GPU.

Metal.device_synchronize — Function

device_synchronize()

Synchronize all committed GPU work across all global queues