Array programming

The Metal array type, MtlArray, generally implements the Base array interface and all of its expected methods.

However, there is the special function mtl for transferring an array over to the gpu. For compatibility reasons, it will automatically convert arrays of Float64 to Float32.

Metal.mtl — Function

mtl(A; storage=Private)

storage can be Private (default) or Shared.

Opinionated GPU array adaptor, which may alter the element type T of arrays:

For T<:AbstractFloat, it makes a MtlArray{Float32} for performance and compatibility reasons (except for Float16).

For T<:Complex{<:AbstractFloat} it makes a MtlArray{ComplexF32}.

For other isbitstype(T), it makes a MtlArray{T}.

By contrast, MtlArray(A) never changes the element type.

Uses Adapt.jl to act inside some wrapper structs.

Examples

julia> mtl(ones(3)')
1×3 adjoint(::MtlVector{Float32, Private}) with eltype Float32:
 1.0  1.0  1.0

julia> mtl(zeros(1,3); storage=Shared)
1×3 MtlMatrix{Float32, Shared}:
 0.0  0.0  0.0

julia> mtl(1:3)
1:3

julia> MtlArray(1:3)
3-element MtlVector{Int64, Private}:
 1
 2
 3

Metal.MtlArray — Type

MtlArray{T,N,S} <: AbstractGPUArray{T,N}

N-dimensional Metal array with storage mode S and elements of type T.

S can be Private (default), Shared, or Managed.

See the Array Programming section of the Metal.jl docs for more details.

Metal.MtlVector — Type

MtlVector{T,S} <: AbstractGPUVector{T}

One-dimensional array with elements of type T for use with Apple Metal-compatible GPUs. Alias for MtlArray{T,1,S}.

See also Vector(@ref), and the Array Programming section of the Metal.jl docs for more details.

Metal.MtlMatrix — Type

MtlMatrix{T,S} <: AbstractGPUMatrix{T}

Two-dimensional array with elements of type T for use with Apple Metal-compatible GPUs. Alias for MtlArray{T,2,S}.

See also Matrix(@ref), and the Array Programming section of the Metal.jl docs for more details.

Metal.MtlVecOrMat — Type

MtlVecOrMat{T,S}

Union type of MtlVector{T,S} and MtlMatrix{T,S} which allows functions to accept either an MtlMatrix or an MtlVector.

See also VecOrMat(@ref) for examples.

Storage modes

The Metal API has various storage modes that dictate how a resource can be accessed. MtlArrays are Private by default, but they can also be Shared or Managed. For more information on storage modes, see the official Metal documentation.

Metal.MTL.Private — Type

struct Private <: MTL.StorageMode

Used to indicate that the resource is stored using MTLStorageModePrivate in memory.

For more information on Metal storage modes, refer to the official Metal documentation.

See also Shared and Managed.

Metal.MTL.Shared — Type

struct Shared <: MTL.StorageMode

Used to indicate that the resource is stored using MTLStorageModeShared in memory.

For more information on Metal storage modes, refer to the official Metal documentation.

See also Private and Managed.

Metal.MTL.Managed — Type

struct Managed <: MTL.StorageMode

Used to indicate that the resource is stored using MTLStorageModeManaged in memory.

For more information on Metal storage modes, refer to the official Metal documentation.

See also Shared and Private.

Metal.MTL.CPUStorage — Type

CPUStorage

Union type of Shared and Managed storage modes.

Represents storage modes where the resource is accessible via the CPU.

There also exist the following convenience functions to check if an MtlArray is using a specific storage mode:

Metal.is_private — Function

is_private(A::MtlArray) -> Bool

Returns true if A has storage mode Private.

See also is_shared and is_managed.

Metal.is_shared — Function

is_shared(A::MtlArray) -> Bool

Returns true if A has storage mode Shared.

See also is_private and is_managed.

Metal.is_managed — Function

is_managed(A::MtlArray) -> Bool

Returns true if A has storage mode Managed.

See also is_shared and is_private.